Attempts to prevent metadataexchange filter from running along with wtf.
Changing proxy config regex to `wtf` doesn't help.
Unclear why this wasn't happening before (doesn't happen when running an
image built a month ago from a343288b, but does happen when checking out
a343288b and rebuilding it. So must be something
in the build - pulling a newer version of an unpinned istio/proxy/envoy
image??). Can't find anything that didn't get committed.

Ended up letting it run (which is fine if it's the original version)